Refactor dto to document, folder name was wrong
