+++ BusinessAssistant-usermanagement/src/main/java/com/businessassistantbcn/usermanagement/document/Role.java
@@ -1,0 +1,6 @@
+package com.businessassistantbcn.usermanagement.document;
+
+public enum Role {
+    USER,
+    ADMIN
+}
