+++ BusinessAssistant-usermanagement/src/main/java/com/businessassistantbcn/usermanagement/document/User.java
@@ -1,4 +1,4 @@
-package com.businessassistantbcn.usermanagement.dto;
+package com.businessassistantbcn.usermanagement.document;
 
 
 import lombok.AllArgsConstructor;
