--- BusinessAssistant-usermanagement/src/main/java/com/businessassistantbcn/usermanagement/dto/Role.java
@@ -1,6 +1,0 @@
-package com.businessassistantbcn.usermanagement.dto;
-
-public enum Role {
-    USER,
-    ADMIN
-}
